#680 - updated integration tests to be more foolproof
