--- test/unit/org/bbop/apollo/FeatureProjectionServiceSpec.groovy
@@ -1,20 +1,0 @@
-package org.bbop.apollo
-
-import grails.test.mixin.TestFor
-import spock.lang.Specification
-
-/**
- * See the API for {@link grails.test.mixin.services.ServiceUnitTestMixin} for usage instructions
- */
-@TestFor(FeatureProjectionService)
-class FeatureProjectionServiceSpec extends Specification {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-    }
-}
